Moving callbacks out of FilterManager and into Explorer
This will allow us to use the FilterManager in other places, specifically for funnel steps.
