+++ client/js/app/components/common/filter_manager.js
@@ -6,17 +6,7 @@ var _ = require('lodash');
 var React = require('react');
 var Filter = require('./filter.js');
 var Modal = require('./modal.js');
-var ExplorerActions = require('../../actions/ExplorerActions');
 var ProjectUtils = require('../../utils/ProjectUtils');
-
-function coerceGeoValue(value) {
-  var trailingDecimals = value.match(/\.+$/);
-  if (value === '-' || (trailingDecimals && trailingDecimals.length)) {
-    return value;
-  } else {
-    return parseFloat(value) || 0;
-  }
-}
 
 var FilterManager = React.createClass({
 
@@ -30,8 +20,8 @@ var FilterManager = React.createClass({
                 propertyType={this.getPropertyType(filter.property_name)}
                 eventCollection={this.props.eventCollection}
                 eventPropertyNames={ProjectUtils.getEventCollectionPropertyNames(this.props.project, this.props.eventCollection)}
-                handleChange={this.handleChange}
-                removeFilter={this.removeFilter}
+                handleChange={this.props.handleChange}
+                removeFilter={this.props.removeFilter}
                 filterOperators={ProjectUtils.getConstant('FILTER_OPERATORS')} />
       );
     }.bind(this));
@@ -40,7 +30,7 @@ var FilterManager = React.createClass({
       <div>
         {filterNodes}
         <div className="filter-buttons">
-          <a href="#" className="add-filter btn btn-primary" onClick={this.addFilter}>
+          <a href="#" className="add-filter btn btn-primary" onClick={this.props.addFilter}>
             <i className="icon glyphicon glyphicon-plus margin-right-tiny"></i>
             Add another filter
           </a>
@@ -64,17 +54,6 @@ var FilterManager = React.createClass({
     );
   },
 
-  addFilter: function(event) {
-    event.preventDefault();
-    ExplorerActions.addFilter(this.props.modelId);
-  },
-
-  removeFilter: function(event) {
-    event.preventDefault();
-    var index = parseInt(event.currentTarget.dataset.index);
-    ExplorerActions.removeFilter(this.props.modelId, index);
-  },
-
   getPropertyType: function (property_name) {
     return ProjectUtils.getPropertyType(
       this.props.project,
@@ -83,32 +62,8 @@ var FilterManager = React.createClass({
     );
   },
 
-  handleChange: function(index, name, value) {
-    var updates = _.cloneDeep(this.props.filters[index]);
-    
-    if (!_.isNull(name.match('coordinates'))) {
-      var coordinateIndex = parseInt(name.split('.')[1]);
-      updates.property_value.coordinates[coordinateIndex] = coerceGeoValue(value);
-    } else if (updates.coercion_type === 'Geo') {
-      updates.property_value[name] = coerceGeoValue(value);
-    } else {
-      updates[name] = value;
-    }
-
-    ExplorerActions.updateFilter(this.props.modelId, index, updates);
-  },
-
-  // React methods
-
-  componentWillMount: function() {
-    // Create a default filter if there are no filters already on this model
-    if (!this.props.filters.length) {
-      ExplorerActions.addFilter(this.props.modelId);
-    }
-  },
-
   render: function() {
-    var filterContent = this.props.eventCollection ? this.buildFilterNodes() : this.noFiltersMarkup()
+    var filterContent = this.props.eventCollection ? this.buildFilterNodes() : this.noFiltersMarkup();
 
     return (
       <Modal ref="modal"
